Spec file to build this cleanly under mock for EL6 systems.
This includes django-admin and ldap functionality and as such has these listed i the depedencies.

At this time the version of django in EPEL is 1.4 and python-django-auth-ldap is not there. The depedencies can be installed via pip (bad way) with the requires removed or a package created via pyp2rpm (or ask for mine) to fullfill the depedency.